+++ src/components/modal/v2/parts/LoadingShimmer.jsx
@@ -1,10 +1,9 @@
 /* eslint-disable eslint-comments/disable-enable-pair */
-/* eslint-disable react/no-array-index-key */
 /** @jsx h */
 import { Fragment, h } from 'preact';
 import arrayFrom from 'core-js-pure/stable/array/from';
 
-const LoadingShimmer = ({ numOffers = 3, offerCountry }) => {
+const LoadingShimmer = ({ numOffers = 3, offerCountry, useNewCheckoutDesign }) => {
     /**
      * Takes the number of offers returned from numOffers and renders the correct
      * number of loading shimmers by creating an array of that size to map over.
@@ -26,7 +25,11 @@ const LoadingShimmer = ({ numOffers = 3, offerCountry }) => {
                     );
                 }
                 return (
-                    <div className="offer__container shimmer" key={index}>
+                    <div
+                        className={`offer__container shimmer key=${index} ${
+                            useNewCheckoutDesign === 'true' ? 'checkout' : ''
+                        }`}
+                    >
                         <div className="offer__row">
                             <div className="offer__field-loading" />
                         </div>