+++ src/components/modal/v2/parts/TermsTable.jsx
@@ -5,7 +5,7 @@ import OfferCard from './OfferCard';
 import OfferAccordion from './OfferAccordion';
 import LoadingShimmer from './LoadingShimmer';
 
-const TermsTable = ({ isLoading, view: { meta, offers }, aprDisclaimer, useV4Design }) => {
+const TermsTable = ({ isLoading, view: { meta, offers }, aprDisclaimer, useV4Design, useNewCheckoutDesign }) => {
     const { offerCountry } = meta;
     /**
      * numOffers/setNumOffers is used to dynamically change the number of loading shimmers that are rendered
@@ -17,7 +17,11 @@ const TermsTable = ({ isLoading, view: { meta, offers }, aprDisclaimer, useV4Des
     if (isLoading) {
         return (
             <div className="offer__wrapper">
-                <LoadingShimmer numOffers={offerCountry !== 'DE' ? numOffers : 4} offerCountry={offerCountry} />
+                <LoadingShimmer
+                    numOffers={offerCountry !== 'DE' ? numOffers : 4}
+                    offerCountry={offerCountry}
+                    useNewCheckoutDesign={useNewCheckoutDesign}
+                />
             </div>
         );
     }
@@ -27,7 +31,14 @@ const TermsTable = ({ isLoading, view: { meta, offers }, aprDisclaimer, useV4Des
         .map((offer, idx) => {
             // Only DE uses the accordion style for presentation of offers in the modal.
             if (offerCountry !== 'DE') {
-                return <OfferCard offer={offer} index={idx} useV4Design={useV4Design} />;
+                return (
+                    <OfferCard
+                        offer={offer}
+                        index={idx}
+                        useV4Design={useV4Design}
+                        useNewCheckoutDesign={useNewCheckoutDesign}
+                    />
+                );
             }
             const disclaimer =
                 aprDisclaimer.length < offers.length