+++ tests/functional/v2/spec/snapshot-tests/api/longTerm.test.js
@@ -50,7 +50,12 @@ descFn.each(filterPermutations([LOCALE_CONFIG], [ACCOUNT]))(
                     getTestName(country, integration, account, amount, viewport)
                 );
             });
-        } else if (amount >= minAmount && amount <= maxAmount && account !== 'DEV_DE_LONG_TERM') {
+        } else if (
+            amount >= minAmount &&
+            amount <= maxAmount &&
+            account !== 'DEV_DE_LONG_TERM' &&
+            account !== 'DEV_DE_LONG_TERM_0APR'
+        ) {
             test(`Amount:${amount} - Offer cards show correct payment headline information - ${viewport}`, async () => {
                 await showCorrectOfferInfo(
                     modalFrame,
@@ -84,6 +89,32 @@ descFn.each(filterPermutations([LOCALE_CONFIG], [ACCOUNT]))(
             });
         }
         if (amount >= minAmount && amount <= maxAmount && account === 'DEV_DE_LONG_TERM') {
+            test(`Amount:${amount} - Offer accordion show correct payment headline information - ${viewport}`, async () => {
+                await showCorrectOfferInfoAccordion(
+                    modalFrame,
+                    modalContent,
+                    getTestName(country, integration, account, amount, viewport)
+                );
+            });
+
+            test(`Amount:${amount} - Offer accordion show correct payment breakdown information for amount - ${viewport}`, async () => {
+                await showCorrectOfferBreakdownAccordion(
+                    modalFrame,
+                    modalContent,
+                    getTestName(country, integration, account, amount, viewport)
+                );
+            });
+
+            test(`Amount:${amount} - Update offers via calculator - ${viewport}`, async () => {
+                await updateOfferAccordionTermsViaCalc(
+                    modalFrame,
+                    modalContent,
+                    getTestName(country, integration, account, amount, viewport)
+                );
+            });
+        }
+
+        if (amount >= minAmount && amount <= maxAmount && account === 'DEV_DE_LONG_TERM_0APR') {
             test(`Amount:${amount} - Offer accordion show correct payment headline information - ${viewport}`, async () => {
                 await showCorrectOfferInfoAccordion(
                     modalFrame,